Хот фикс по модели Контактов
