+++ backend/contacts/apps.py
@@ -7,4 +7,4 @@ class ContactsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "contacts"
     verbose_name = "Контакт"
-    verbose_name_pliral = "Контакты"
+    verbose_name_plural = "Контакты"
